hw3/requirements.txt
functions-framework==3.*
google-cloud-pubsub==2.18.0
grpcio==1.59.0
google-cloud-storage
flask
logging